requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
asgiref=3.2.7=pypi_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py36_1
django=3.0.4=pypi_0
django-crispy-forms=1.9.0=pypi_0
django-formtools=2.2=pypi_0
django-widget-tweaks=1.4.8=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
ncurses=6.2=he6710b0_0
openssl=1.1.1e=h7b6447c_0
pip=20.0.2=py36_1
python=3.6.10=hcf32534_1
pytz=2019.3=pypi_0
readline=8.0=h7b6447c_0
setuptools=46.1.1=py36_0
sqlite=3.31.1=h7b6447c_0
sqlparse=0.3.1=pypi_0
tk=8.6.8=hbc83047_0
wheel=0.34.2=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
